Extrat renderTooltip function to utils
