+++ packages/visualizations/src/components/Map/WebGl/ChoroplethGeoJson.svelte
@@ -2,13 +2,13 @@
 
 <script lang="ts">
     import turfBbox from '@turf/bbox';
-    import { debounce } from 'lodash';
     import type { SourceSpecification } from 'maplibre-gl';
     import type { BBox } from 'geojson';
+    import { debounce } from 'lodash';
     import type { ColorScale, DataBounds, Color } from '../../types';
     import MapRender from './MapRender.svelte';
     import { BLANK } from '../mapStyles';
-    import { getDataBounds, mapKeyToColor, VOID_BOUNDS, computeFilterExpression, defaultFormat } from '../utils';
+    import { getDataBounds, mapKeyToColor, VOID_BOUNDS, computeFilterExpression, computeTooltip } from '../utils';
     import { DEFAULT_COLORS, DEFAULT_COLORS_SCALE } from '../constants';
     import type {
         ChoroplethDataValue,
@@ -65,12 +65,6 @@
         newColorScale: ColorScale,
         values: ChoroplethDataValue[] = []
     ) {
-        // Specific to move outside ? On Call ?
-        if (!newShapes.geoJson) {
-            // We don't have everything we need yet
-            return;
-        }
-
         let colors;
         let fillColor: string | (string | string[])[] = emptyValueColor;
 
@@ -92,58 +86,20 @@
                 'fill-outline-color': DEFAULT_COLORS.ShapeOutline,
             },
         };
-        // Specific
-        if (newShapes.geoJson) {
-            source = {
-                type: 'geojson',
-                data: newShapes.geoJson,
-            };
+        source = {
+            type: 'geojson',
+            data: newShapes.geoJson,
+        };
 
-            layer = baseLayer;
-            // Specific
-            bbox = bbox || turfBbox(newShapes.geoJson) || VOID_BOUNDS;
-        }
+        layer = baseLayer;
+
+        bbox = bbox || turfBbox(newShapes.geoJson) || VOID_BOUNDS;
     }
 
-    $: computeSourceLayerAndBboxes(shapes, colorScale, data.value);
+    $: if (shapes.geoJson) {computeSourceLayerAndBboxes(shapes, colorScale, data.value);}
 
     $: renderTooltip = debounce(
-        (hoveredFeature) => {
-            const values = data.value || [];
-            const matchedFeature = values.find(
-                (item) => String(item.x) === hoveredFeature.properties?.[matchKey]
-            );
-
-            let tooltipLabel =
-                hoveredFeature.properties?.label || hoveredFeature.properties?.[matchKey];
-            const labelMatcher = options?.tooltip?.labelMatcher;
-
-            if (labelMatcher && matchedFeature) {
-                const { type } = labelMatcher;
-                if (type === 'keyProperty') {
-                    const { key } = labelMatcher;
-                    tooltipLabel = hoveredFeature.properties?.[key];
-                } else if (type === 'keyMap') {
-                    const { mapping } = labelMatcher;
-                    tooltipLabel = mapping[matchedFeature?.x];
-                }
-            }
-
-            const tooltipRawValues: {
-                value?: number;
-                label: string;
-                key: string;
-            } = {
-                value: matchedFeature?.y,
-                label: tooltipLabel,
-                key: hoveredFeature.properties?.[matchKey], // === matchedFeature.x
-            };
-            const format = options?.tooltip?.labelFormatter;
-
-            return format ? format(tooltipRawValues) : defaultFormat(tooltipRawValues);
-        },
-        10,
-        { leading: true }
+        (hoveredFeature) => computeTooltip(hoveredFeature, data.value, options, matchKey), 10, {leading: true}
     );
 
     $: if (filter) {
