+++ packages/visualizations/src/components/Map/WebGl/ChoroplethVectorTiles.svelte
@@ -1,22 +1,22 @@
 <svelte:options immutable={true} />
 
 <script lang="ts">
-    import { debounce } from 'lodash';
     import type { SourceSpecification } from 'maplibre-gl';
     import type { BBox } from 'geojson';
+    import { debounce } from 'lodash';
     import type { ColorScale, DataBounds, Color } from '../../types';
     import MapRender from './MapRender.svelte';
     import { BLANK } from '../mapStyles';
-    import { getDataBounds, mapKeyToColor, VOID_BOUNDS, computeFilterExpression, defaultFormat } from '../utils';
+    import { getDataBounds, mapKeyToColor, VOID_BOUNDS, computeFilterExpression, computeTooltip } from '../utils';
     import { DEFAULT_COLORS, DEFAULT_COLORS_SCALE } from '../constants';
     import type {
         ChoroplethDataValue,
         ChoroplethLayer,
         VectorChoroplethOptions,
-        MapRenderTooltipFunction,
         ChoroplethShapeVectorTilesValue,
         MapLegend,
         MapFilter,
+        MapRenderTooltipFunction,
     } from '../types';
 
     export let data: { value: ChoroplethDataValue[] }; // values, and the key to match
@@ -66,12 +66,6 @@
         newColorScale: ColorScale,
         values: ChoroplethDataValue[] = []
     ) {
-        // Specific to move outside ? on Call ?
-        if (!newShapes.url) {
-            // We don't have everything we need yet
-            return;
-        }
-
         let colors;
         let fillColor: string | (string | string[])[] = emptyValueColor;
 
@@ -93,7 +87,7 @@
                 'fill-outline-color': DEFAULT_COLORS.ShapeOutline,
             },
         };
-        // Specific
+
         source = {
             type: 'vector',
             tiles: [newShapes.url],
@@ -103,49 +97,14 @@
             ...baseLayer,
             'source-layer': newShapes.layer,
         };
-        // Specific
+
         bbox = bbox || VOID_BOUNDS;
     }
 
-    $: computeSourceLayerAndBboxes(shapes, colorScale, data.value);
+    $: if (shapes.url) {computeSourceLayerAndBboxes(shapes, colorScale, data.value);}
 
     $: renderTooltip = debounce(
-        (hoveredFeature) => {
-            const values = data.value || [];
-            const matchedFeature = values.find(
-                (item) => String(item.x) === hoveredFeature.properties?.[matchKey]
-            );
-
-            let tooltipLabel =
-                hoveredFeature.properties?.label || hoveredFeature.properties?.[matchKey];
-            const labelMatcher = options?.tooltip?.labelMatcher;
-
-            if (labelMatcher && matchedFeature) {
-                const { type } = labelMatcher;
-                if (type === 'keyProperty') {
-                    const { key } = labelMatcher;
-                    tooltipLabel = hoveredFeature.properties?.[key];
-                } else if (type === 'keyMap') {
-                    const { mapping } = labelMatcher;
-                    tooltipLabel = mapping[matchedFeature?.x];
-                }
-            }
-
-            const tooltipRawValues: {
-                value?: number;
-                label: string;
-                key: string;
-            } = {
-                value: matchedFeature?.y,
-                label: tooltipLabel,
-                key: hoveredFeature.properties?.[matchKey], // === matchedFeature.x
-            };
-            const format = options?.tooltip?.labelFormatter;
-
-            return format ? format(tooltipRawValues) : defaultFormat(tooltipRawValues);
-        },
-        10,
-        { leading: true }
+        (hoveredFeature) => computeTooltip(hoveredFeature, data.value, options, matchKey), 10, {leading: true}
     );
 
     $: if (filter) {
